+++ src/lnav_config.hh
@@ -80,6 +80,7 @@ bool update_installs_from_git();
 void install_extra_formats();
 
 struct key_command {
+    std::string kc_id;
     std::string kc_cmd;
     std::string kc_alt_msg;
 };